requirements.txt
attrs==23.2.0
certifi==2023.11.17
contourpy==1.2.0
cycler==0.12.1
dnspython==2.6.1
exceptiongroup==1.2.0
fonttools==4.49.0
h11==0.14.0
idna==3.6
kiwisolver==1.4.5
matplotlib==3.8.3
numpy==1.26.4
outcome==1.3.0.post0
packaging==23.2
pathlib==1.0.1
pillow==10.2.0
pymongo==4.6.2
pyparsing==3.1.2
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.24.0
trio-websocket==0.11.1
urllib3==2.1.0
wsproto==1.2.0
